requirements.txt
fastapi[all]==0.78.0
numpy==1.22.4
pathlib==1.0.1
torch==2.1.0
torchvision==0.16.0
efficientnet-pytorch==0.7.1
Pillow==9.4.0
requests==2.26.0
uvicorn[standard]==0.17.6
gunicorn==21.2.0
